Refactor color patches example
