+++ examples/ColorPatches/run.py
@@ -1,0 +1,3 @@
+from color_patches.server import server
+
+server.launch()
